% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/addAvgPsuDepth.R
\name{addAvgPsuDepth}
\alias{addAvgPsuDepth}
\title{Adds AVERAGE_PSU_DEPTH to a data.frame}
\usage{
addAvgPsuDepth(x)
}
\arguments{
\item{x}{A data.frame containing the columns DEPTH, STATION_NR,
and PRIMARY_SAMPLE_UNIT}
}
\value{
a data.frame with AVERAGE_PSU_DEPTH column added
}
\description{
Adds AVERAGE_PSU_DEPTH to a data.frame
}

