% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/addProtectedStatus.R
\name{addProtectedStatus}
\alias{addProtectedStatus}
\title{Generates a column of protected statuses}
\usage{
addProtectedStatus(x)
}
\arguments{
\item{x}{A data.frame containing AR1.0 data which
contains columns called MPA_NR, and SUBREGION_NR}
}
\value{
A data.frame of the original data with a PROT
column added
}
\description{
Generates a column, PROT, with the protected
status of each record given a data.frame of
AR1.0 data
}

