% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/addStrata.R
\name{addStrata}
\alias{addStrata}
\title{Adds strata to AR1.0 data}
\usage{
addStrata(x, region)
}
\arguments{
\item{x}{A data.frame containing AR1.0 data with a
column PROT indicating protected status and a
column HABITAT_CD indicating the habitat}

\item{region}{A string of the region the AR1.0 data is collected from:
FLA KEYS, SEFCRI, or DRTO}
}
\value{
A data.frame with the original AR2.0 data with
a STRAT column added containing the stratum to
which a record belongs
}
\description{
Adds a column of strata to AR1.0 data
}

