% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/addStrataDT.R
\name{addStrataDT}
\alias{addStrataDT}
\title{Adds strata for Dry Tortugas region}
\usage{
addStrataDT(x)
}
\arguments{
\item{x}{A data.frame containing AR1.0 data with a
column PROT indicating protected status and a
column HABITAT_CD indicating the habitat}
}
\value{
A data.frame with the original AR2.0 data with
a STRAT and REGION column added
}
\description{
Adds strata for Dry Tortugas region
}

