% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analysisready.R
\docType{package}
\name{analysisready}
\alias{analysisready}
\alias{analysisready-package}
\title{A package to convert AR1.0 data to AR2.0 data}
\description{
Takes csv files of AR1.0 data, drops data not usable in analysis,
and adds columns for REGION, STRAT, and PROT
}
\author{
Harry Ganz
}

