% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dropInvalidHabitats.R
\name{dropInvalidHabitats}
\alias{dropInvalidHabitats}
\title{Drops records with invalid habitats}
\usage{
dropInvalidHabitats(x, invalid_habs = c("ARFT_NA", "SAND_NA", "SGRS_NA",
  "UCHB_LR"))
}
\arguments{
\item{x}{A data.frame containing AR1.0 data
with a column named HABITAT_CD}

\item{invalid_habs}{A character vector of all the invalid habitat
codes}
}
\value{
A data.frame with records
from invalid habitats removed
}
\description{
Drops records with invalid habitats
}

