% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dropTimeSeen.R
\name{dropTimeSeen}
\alias{dropTimeSeen}
\title{Drops rows from AR1.0 data based on TIME_SEEN}
\usage{
dropTimeSeen(x, lookup = NULL)
}
\arguments{
\item{x}{A data.frame containing AR1.0 data with, at least,
a column named SPECIES_CD containing species codes
and a column named TIME_SEEN containing the times
in which the data was recorded}

\item{lookup}{A data.frame containing a column named SPECIES_CD
with species codes and KEEP_10MIN containing
a 1 if the species should be included after 10
minutes (TIME_SEEN > 2) and a zero otherwise.
If NULL function just drops all records
with TIME_SEEN > 2}
}
\value{
A data.frame of the AR1.0 data with
records collected after 10 minutes or
whatever was specified in the lookup table
 dropped
}
\description{
Drops records from AR1.0 data with TIME_SEEN
> 2 or based on information from a lookup table
}

