% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dropUnderwaterVisibility.R
\name{dropUnderwaterVisibility}
\alias{dropUnderwaterVisibility}
\title{Drops rows from data based on visibility}
\usage{
dropUnderwaterVisibility(x, cutoff = 3)
}
\arguments{
\item{x}{A data.frame containing AR1.0 data with a column
called UNDERWATER_VISIBILITY}

\item{cutoff}{A numeric value of visibility (in meters)
below which to drop records. Default
value is 3 meters}
}
\value{
A data.frame of AR1.0 data subset to exclude
visibility values below the cutoff
}
\description{
Drop rows from AR1.0 data with
UNDERWATER_VISIBILITY below cutoff value and
return a data.frame
}

