% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/toAr2.R
\name{toAr2}
\alias{toAr2}
\title{Converts AR1.0 data to AR2.0 data}
\usage{
toAr2(x, region, lookup = NULL, cutoff = 3)
}
\arguments{
\item{x}{a data.frame of AR1.0 data}
}
\value{
A data.frame of AR2.0 data
}
\description{
Takes AR1.0 data and converts it to AR2.0 data
}

